requirements.txt
selenium==4.15.0
pytest==7.4.0
pytest-xdist==3.3.1
allure-pytest==2.13.2
webdriver-manager==4.0.1
pytest-html==3.2.0
requests==2.31.0
pydantic[email]==2.5.0
faker==19.6.2